double -> floats

+++ libs2/GeOS.hpp
@@ -12,11 +12,11 @@ class Matrix;
 
 class GeomObj {
 protected:
-  double g[4];
+  float g[4];
   GeomObj(){
     for (int i = 0; i < 4; i++) g[i] = 0;
   }
-  GeomObj(int xi, int yj, int zk){
+  GeomObj(float xi, float yj, float zk){
     g[0] = xi; g[1] = yj; g[2] = zk;
   }
   friend class Matrix;
@@ -27,7 +27,7 @@ public:
   bool isDirection() const {
     return g[3] == 0;
   }
-  double& operator[](GeomComponent c){
+  float& operator[](GeomComponent c){
     switch (c) {
       case xi:
         return g[0];
@@ -37,7 +37,7 @@ public:
         return g[2];
     }
   }
-  double operator[](GeomComponent c) const {
+  float operator[](GeomComponent c) const {
     switch (c) {
       case xi:
         return g[0];
@@ -47,16 +47,16 @@ public:
         return g[2];
     }
   }
-  double getxi() const {
+  float getxi() const {
     return g[0];
   }
-  double getyj() const {
+  float getyj() const {
     return g[1];
   }
-  double getzk() const {
+  float getzk() const {
     return g[2];
   }
-  double getType() const {
+  float getType() const {
     return g[3];
   }
   void setxi(int xi){
@@ -75,7 +75,7 @@ public:
   Direction() : GeomObj() {
     g[3] = 0;
   }
-  Direction(int x, int y, int z) : GeomObj(x, y, z) {
+  Direction(float x, float y, float z) : GeomObj(x, y, z) {
     g[3] = 0;
   }
 
@@ -85,14 +85,14 @@ public:
   Direction operator+(const Direction& other) const {
     return Direction(g[0] + other.g[0], g[1] + other.g[1], g[2] + other.g[2]);
   }
-  Direction operator*(double escalar) const {
+  Direction operator*(float escalar) const {
     return Direction(g[0]*escalar, g[1]*escalar, g[2]*escalar);
   }
-  double modulus() const {
+  float modulus() const {
     return sqrt( g[0]*g[0] + g[1]*g[1] + g[2]*g[2]);
   }
   Direction normalize() {
-    double mod = modulus();
+    float mod = modulus();
     if (mod != 0){
       g[0] /= mod;
       g[1] /= mod;
@@ -107,7 +107,7 @@ public:
   Point() : GeomObj() {
     g[3] = 1;
   }
-  Point(int x, int y, int z) : GeomObj(x, y, z) {
+  Point(float x, float y, float z) : GeomObj(x, y, z) {
     g[3] = 1;
   }
 
@@ -121,11 +121,11 @@ public:
 
 
 
-double dotProduct(const Direction& a, const Direction& b) {
+float dotProduct(const Direction& a, const Direction& b) {
   return a[xi]*b[xi] + a[yj]*b[yj] + a[zk]*b[zk];
 }
 Direction crossProduct(const Direction& a, const Direction& b) {
-  double i = a[yj]*b[zk] - a[zk]*b[yj],
+  float i = a[yj]*b[zk] - a[zk]*b[yj],
          j = a[zk]*b[xi] - a[xi]*b[zk],
          k = a[xi]*b[yj] - a[yj]*b[xi];
   return Direction(i,j,k);
@@ -155,7 +155,7 @@ protected:
   RGB solid_color;
   Object(){};
 public:
-  virtual bool intersection(const Direction& ray, const Point& origen, double &dist) = 0;
+  virtual bool intersection(const Direction& ray, const Point& origen, float &dist) = 0;
 
   void setRGB(RGB sc){
     solid_color = sc;
